% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditch_mac_cruft.R
\name{ditch_mac_cruft}
\alias{ditch_mac_cruft}
\title{Ditch Mac cruft}
\usage{
ditch_mac_cruft()
}
\value{

}
\description{
Drops all those annoying Mac-specific files such as \code{.DS_Store} within the
current R project directory and subdirectories.
}
