% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up common packages and directories}
\usage{
setup(dirs = TRUE)
}
\arguments{
\item{dirs}{Should paths to Mazu be loaded? default \code{TRUE}}
}
\value{
nothing to return
}
\description{
Attach \code{tidyverse}, \code{here}, \code{RColorBrewer} packages; create aliases to locations
on Mazu (optional)
}
