% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ohara.R
\name{theme_ohara}
\alias{theme_ohara}
\title{A theme for `ggplot`}
\usage{
theme_ohara(base_size = 9)
}
\arguments{
\item{base_size}{}
}
\value{
`ggplot` `theme` object
}
\description{
A theme for `ggplot`
}
