% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_join.R
\name{dt_join}
\alias{dt_join}
\title{Join dataframes using datatable functionality}
\usage{
dt_join(df1, df2, by, type)
}
\arguments{
\item{df1}{a large dataframe}

\item{df2}{a large dataframe}

\item{by}{a column to be used as a join key.  Column name must be same in both `df1` and `df2`.}

\item{type}{What type of join? options 'left', 'full', 'inner'.}
}
\value{
A joined dataframe
}
\description{
A join function for very large dataframes with similar syntax to `dplyr::left_join`,
but using faster `data.table::merge()`.
}
