% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dupes.R
\name{show_dupes}
\alias{show_dupes}
\title{Show duplicate items in a dataframe}
\usage{
show_dupes(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{A dataframe}

\item{y}{A character column name within dataframe \code{x}}

\item{na.rm}{}
}
\value{
A dataframe that includes all rows of \code{x} with duplicated values
in the given column \code{y}.
}
\description{
Identifies and returns duplicated values within a given column of a dataframe.
}
